% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_asset.R
\name{calc_omee}
\alias{calc_omee}
\title{Overall Mining Equipment Effectiveness}
\usage{
calc_omee(availability, utilization, speed, bucket)
}
\arguments{
\item{availability}{numeric}

\item{utilization}{numeric}

\item{speed}{numeric}

\item{bucket}{numeric}
}
\value{
numeric
}
\description{
Calculates mining OMEE.
}
