% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kpis.R
\name{list_kpi_categories}
\alias{list_kpi_categories}
\title{List KPI Categories}
\usage{
list_kpi_categories()
}
\value{
character vector
}
\description{
Returns all KPI categories available in MiningAnalytics.
}
