% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mine_esg_score.R
\name{mine_esg_score}
\alias{mine_esg_score}
\title{Mine ESG Score}
\usage{
mine_esg_score(environmental, social, governance)
}
\arguments{
\item{environmental}{environmental score}

\item{social}{social score}

\item{governance}{governance score}
}
\value{
ESG score
}
\description{
Calculate ESG performance score
}
